\name{shinyInfl}
\alias{shinyInfl}
\encoding{latin1}

\title{Initializes a Shiny App with all implemented methods of this package}

\description{
A comprehensive Shiny App that facilitates the import of user-supplied data and subsequent detailed testing of all implemented methods in this package. Analysis results can be exported as plots and text. 
}

\usage{
shinyInfl() 
}

\arguments{
None.
}

\value{
The analysis including plots and result tables.
}

\author{
Andrej-Nikolai Spiess
}   

\examples{
## shinyInfl() # <= to initialize
}   

\keyword{optimize}
\keyword{models}
\keyword{linear}

